// models/chest.ts
/**
 * A list of items that can be obtained from opening a chest.
 */
export enum ChestItem {
    /** food items */
    APPLE = 'Apple',
    CHOCOLATE = 'Chocolate',
    JUICE = 'Juice',
    BURGER = 'Burger',
    /** resources */
    SEAWEED = 'Seaweed',
    STONE = 'STONE',
    KERATIN = 'KERATIN',
    SILVER = 'SILVER',
    DIAMOND = 'DIAMOND',
    RELIC = 'Relic',
    /** rest of items */
    X_COOKIES = 'xCookies',
    TERRA_CAPSULATOR = 'Terra Capsulator',
    BIT_ORB = 'Bit Orb',
}